# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: FIS Process API
  version: v1
  description: Complete REST API contract for the FIS ledger service.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /accounts:
    post:
      summary: Create account
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ActorIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
    get:
      summary: List accounts
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: accountType
          in: query
          schema:
            type: string
            enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Account page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPage'
  /accounts/{accountCode}:
    get:
      summary: Get account by code
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: accountCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update account
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ActorIdHeader'
        - name: accountCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
  /accounts/{accountCode}/aggregated-balance:
    get:
      summary: Get account aggregated balance (self + descendants)
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - name: accountCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account with aggregated balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/integrity-check:
    get:
      summary: Verify accounting equation integrity for a tenant
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Integrity check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerIntegrityCheckResponse'
  /admin/year-end-close:
    post:
      summary: Perform fiscal year-end close
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YearEndCloseRequest'
      responses:
        '200':
          description: Year-end close result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YearEndCloseResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
    ActorIdHeader:
      name: X-Actor-Id
      in: header
      required: false
      schema:
        type: string
...
  responses:
    ValidationFailed:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Entity not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    UnprocessableEntity:
      description: Business rule rejection
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    CreateAccountRequest:
      type: object
      required: [code, name, accountType, currencyCode]
      properties:
        code:
          type: string
        name:
          type: string
        accountType:
          type: string
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        currencyCode:
          type: string
        parentAccountCode:
          type: string
          nullable: true
        contra:
          type: boolean
    UpdateAccountRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        isActive:
          type: boolean
          nullable: true
    AccountResponse:
      type: object
      properties:
        accountId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        accountType:
          type: string
        currencyCode:
          type: string
        currentBalanceCents:
          type: integer
          format: int64
        formattedBalance:
          type: string
        isActive:
          type: boolean
        contra:
          type: boolean
        parentAccountCode:
          type: string
          nullable: true
        aggregatedBalanceCents:
          type: integer
          format: int64
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AccountPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...
    LedgerIntegrityCheckResponse:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        assetTotal:
          type: integer
          format: int64
        liabilityTotal:
          type: integer
          format: int64
        equityTotal:
          type: integer
          format: int64
        revenueTotal:
          type: integer
          format: int64
        expenseTotal:
          type: integer
          format: int64
        equationDelta:
          type: integer
          format: int64
        balanced:
          type: boolean
    YearEndCloseRequest:
      type: object
      required: [fiscalYear, retainedEarningsAccountCode, createdBy]
      properties:
        fiscalYear:
          type: integer
        retainedEarningsAccountCode:
          type: string
        createdBy:
          type: string
    YearEndCloseResponse:
      type: object
      properties:
        fiscalYear:
          type: integer
        closingJournalEntryId:
          type: string
          format: uuid
          nullable: true
        totalRevenue:
          type: integer
          format: int64
        totalExpenses:
          type: integer
          format: int64
        netIncome:
          type: integer
          format: int64
        accountsClosed:
          type: integer
        retainedEarningsAccountCode:
          type: string
        message:
          type: string
...